update fitur setting,dashboard,and fix bug

+++ app.py
@@ -19,13 +19,34 @@ def get_db():
 conn = get_db()
 
 users = {}
+def get_user_data(user_id):
+    cursor = conn.cursor()
+    sql_query = "SELECT name, username FROM users WHERE id = %s"
+    cursor.execute(sql_query, (user_id,))
+    result = cursor.fetchone()
 
+    cursor.close()
+    return result 
+@app.context_processor
+def inject_user_data():
+    user_id = session.get("user_id", None)
+    user_data = get_user_data(user_id)
+
+    return dict(user_id=user_id, user_data=user_data)
 
 class User:
     def __init__(self, id, name, username):
         self.id = id
         self.name = name
         self.username = username
+
+    def update(self):
+        cursor = conn.cursor()
+        sql_update_income = """UPDATE users SET name=%s, username=%s WHERE id=%s"""
+        cursor.execute(sql_update_income, (self.name, self.username, self.id))
+        conn.commit()
+
+        cursor.close()
 
 class Income:
     def __init__(self, user_id, nominal, source_type, note, date, id=None):
@@ -38,22 +59,68 @@ class Income:
 
     def save(self):
         cursor = conn.cursor()
-        sql = """INSERT INTO income (user_id, nominal, source_type, note, date) VALUES (%s, %s, %s, %s, %s)"""
-        cursor.execute(sql, (self.user_id, self.nominal, self.source_type, self.note, self.date))
+        sql_insert_income = """INSERT INTO income (user_id, nominal, source_type, note, date) VALUES (%s, %s, %s, %s, %s)"""
+        cursor.execute(sql_insert_income, (self.user_id, self.nominal, self.source_type, self.note, self.date))
         conn.commit()
+
+        sql_update_saldo = """UPDATE users SET saldo = saldo + %s WHERE id = %s"""
+        cursor.execute(sql_update_saldo, (self.nominal, self.user_id))
+        conn.commit()
+
+        cursor.close()
+
         
     def update(self):
         cursor = conn.cursor()
-        sql = """UPDATE income SET user_id=%s, nominal=%s, source_type=%s, note=%s, date=%s WHERE id=%s"""
-        cursor.execute(sql, (self.user_id, self.nominal, self.source_type, self.note, self.date, self.id))
-        conn.commit()
+        sql_get_initial_nominal = """SELECT nominal, user_id FROM income WHERE id=%s"""
+        cursor.execute(sql_get_initial_nominal, (self.id,))
+        result = cursor.fetchone()
+
+        if result:
+            initial_nominal = int(result[0])  # Convert to int
+            user_id = result[1]
+
+            new_nominal = int(self.nominal) if isinstance(self.nominal, str) else self.nominal
+            sql_update_income = """UPDATE income SET user_id=%s, nominal=%s, source_type=%s, note=%s, date=%s WHERE id=%s"""
+            cursor.execute(sql_update_income, (self.user_id, new_nominal, self.source_type, self.note, self.date, self.id))
+            conn.commit()
+
+            nominal_difference = new_nominal - initial_nominal
+
+            sql_update_saldo = """UPDATE users SET saldo = saldo + %s WHERE id = %s"""
+            cursor.execute(sql_update_saldo, (nominal_difference, user_id))
+            conn.commit()
+        else:
+            print("Income data not found")
+
+        cursor.close()
+
+
 
     def delete(self, id):
         cursor = conn.cursor()
-        sql = """DELETE FROM income WHERE id=%s"""
-        cursor.execute(sql, (id,))
-        conn.commit()
+        sql_get_nominal = """SELECT nominal, user_id FROM income WHERE id=%s"""
+        cursor.execute(sql_get_nominal, (id,))
+        result = cursor.fetchone()
+
+        if result:
+            nominal = result[0]
+            user_id = result[1]
+
+            # Delete income data
+            sql_delete_income = """DELETE FROM income WHERE id=%s"""
+            cursor.execute(sql_delete_income, (id,))
+            conn.commit()
+
+            # Update user saldo
+            sql_update_saldo = """UPDATE users SET saldo = saldo - %s WHERE id = %s"""
+            cursor.execute(sql_update_saldo, (nominal, user_id))
+            conn.commit()
+        else:
+            print("Income data not found")
+
         cursor.close()
+
     
 
 class Outcome:
@@ -70,18 +137,56 @@ class Outcome:
         sql = """INSERT INTO outcome (user_id, nominal, used_type, note, date) VALUES (%s, %s, %s, %s, %s)"""
         cursor.execute(sql, (self.user_id, self.nominal, self.used_type, self.note, self.date))
         conn.commit()
+
+        sql_update_saldo = """UPDATE users SET saldo = saldo - %s WHERE id = %s"""
+        cursor.execute(sql_update_saldo, (self.nominal, self.user_id))
+        conn.commit()
+        cursor.close()
         
     def update(self):
         cursor = conn.cursor()
-        sql = """UPDATE outcome SET user_id=%s, nominal=%s, used_type=%s, note=%s, date=%s WHERE id=%s"""
-        cursor.execute(sql, (self.user_id, self.nominal, self.used_type, self.note, self.date, self.id))
-        conn.commit()
+        sql_get_initial_nominal = """SELECT nominal, user_id FROM outcome WHERE id=%s"""
+        cursor.execute(sql_get_initial_nominal, (self.id,))
+        result = cursor.fetchone()
+
+        if result:
+            initial_nominal = int(result[0])  # Convert to int
+            user_id = result[1]
+
+            new_nominal = int(self.nominal) if isinstance(self.nominal, str) else self.nominal
+            sql = """UPDATE outcome SET user_id=%s, nominal=%s, used_type=%s, note=%s, date=%s WHERE id=%s"""
+            cursor.execute(sql, (self.user_id, self.nominal, self.used_type, self.note, self.date, self.id))
+            conn.commit()
+            nominal_difference = new_nominal - initial_nominal
+
+            sql_update_saldo = """UPDATE users SET saldo = saldo - %s WHERE id = %s"""
+            cursor.execute(sql_update_saldo, (nominal_difference, user_id))
+            conn.commit()
+        else:
+            print("Outcome data not found")
+
+        cursor.close()
 
     def delete(self, id):
         cursor = conn.cursor()
-        sql = """DELETE FROM outcome WHERE id=%s"""
-        cursor.execute(sql, (id,))
-        conn.commit()
+        sql_get_nominal = """SELECT nominal, user_id FROM outcome WHERE id=%s"""
+        cursor.execute(sql_get_nominal, (id,))
+        result = cursor.fetchone()
+
+        if result:
+            nominal = result[0]
+            user_id = result[1]
+
+            sql_delete_outcome = """DELETE FROM outcome WHERE id=%s"""
+            cursor.execute(sql_delete_outcome, (id,))
+            conn.commit()
+
+            sql_update_saldo = """UPDATE users SET saldo = saldo + %s WHERE id = %s"""
+            cursor.execute(sql_update_saldo, (nominal, user_id))
+            conn.commit()
+        else:
+            print("outcome data not found")
+
         cursor.close()
     
 def login_required(func):
@@ -98,10 +203,16 @@ def index():
     conn = get_db()
     cursor = conn.cursor()
     user_id = session["user_id"]
-    cursor.execute("SELECT * FROM users WHERE id=%s", (user_id,))
-    user = cursor.fetchone()
-    dict_data = dict(zip([column[0] for column in cursor.description], user))
-    return render_template('dashboard.html', data=dict_data)
+
+    sql_query = """ SELECT users.id, users.username, users.saldo,( SELECT COALESCE(SUM(nominal), 0) FROM income WHERE user_id = users.id) AS total_income,( SELECT COALESCE(SUM(nominal), 0) FROM outcome WHERE user_id = users.id) AS total_outcome FROM users WHERE users.id = %s """
+
+    cursor.execute(sql_query, (user_id,))
+    result = cursor.fetchone()
+
+    if result:
+        dict_data = dict(zip([column[0] for column in cursor.description], result))
+        return render_template('dashboard.html', data=dict_data)
+    
 
 # INCOME ROUTE
 
@@ -119,45 +230,55 @@ def income():
 @login_required
 def create_income():
     if request.method == "POST":
-        user_id = session["user_id"]
-        nominal = request.form["nominal"]
-        source_type = request.form["source_type"]
-        note = request.form["note"]
-        date = request.form["date"]
+        try :
+            user_id = session["user_id"]
+            nominal = request.form["nominal"]
+            source_type = request.form["source_type"]
+            note = request.form["note"]
+            date = request.form["date"]
 
-        income = Income(user_id=user_id, nominal=nominal, source_type=source_type, note=note, date=date)
-        income.save()
-
-        # flash('Data added successfully', 'success')
-        return redirect(url_for('income'))
+            income = Income(user_id=user_id, nominal=nominal, source_type=source_type, note=note, date=date)
+            income.save()
+            flash('Data added successfully', 'success')
+            return redirect(url_for('income'))
+        except Exception as e:
+            print(f"Error added data: {e}")
+            flash('Error added data', 'error')
     return render_template('income.html')
 
 @app.route("/income/update", methods=['GET', 'POST'])
 def update_income():
     if request.method == "POST":
-        user_id = session["user_id"]
-        nominal = request.form["nominal"]
-        source_type = request.form["source_type"]
-        note = request.form["note"]
-        date = request.form["date"]
-        id = request.form["id"]
+        try :
+            user_id = session["user_id"]
+            nominal = request.form["nominal"]
+            source_type = request.form["source_type"]
+            note = request.form["note"]
+            date = request.form["date"]
+            id = request.form["id"]
 
-        income = Income(user_id=user_id, nominal=nominal, source_type=source_type, note=note, date=date, id=id)
-        income.update()
-
-        flash('Data update successfully', 'success')
-        return redirect(url_for('income'))
+            income = Income(user_id=user_id, nominal=nominal, source_type=source_type, note=note, date=date, id=id)
+            income.update()
+            flash('Data update successfully', 'success')
+            return redirect(url_for('income'))
+        except Exception as e:
+            print(f"Error updating data: {e}")
+            flash('Error updating data', 'error')
     return render_template('income.html')
 
 @app.route("/income/<income_id>/delete", methods=['GET', 'POST'])
 def delete_income(income_id):
     print(income_id)   
     if request.method == 'GET':
-        income = Income(None, None, None, None, None, None)
-        income.delete(income_id)
+        try :
+            income = Income(None, None, None, None, None, None)
+            income.delete(income_id)
         
-        flash('Data deleted successfully', 'success')
-        return redirect(url_for('income'))
+            flash('Data deleted successfully', 'success')
+            return redirect(url_for('income'))
+        except Exception as e:
+            print(f"Error deleted data: {e}")
+            flash('Error deleted data', 'error')
     return render_template('income.html')
 
 # OUTCOME ROUTE
@@ -177,53 +298,57 @@ def outcome():
 @login_required
 def create_outcome():
     if request.method == "POST":
-        user_id = session["user_id"]
-        nominal = request.form["nominal"]
-        used_type = request.form["used_type"]
-        note = request.form["note"]
-        date = request.form["date"]
+        try :
+            user_id = session["user_id"]
+            nominal = request.form["nominal"]
+            used_type = request.form["used_type"]
+            note = request.form["note"]
+            date = request.form["date"]
 
-        outcome = Outcome(user_id=user_id, nominal=nominal, used_type=used_type, note=note, date=date)
-        outcome.save()
-
-        # flash('Data added successfully', 'success')
-        return redirect(url_for('outcome'))
+            outcome = Outcome(user_id=user_id, nominal=nominal, used_type=used_type, note=note, date=date)
+            outcome.save()
+            flash('Data added successfully', 'success')
+            return redirect(url_for('outcome'))
+        except Exception as e:
+            print(f"Error added data: {e}")
+            flash('Error added data', 'error')
+            
     return render_template('outcome.html')
 
 @app.route("/outcome/update", methods=['GET', 'POST'])
 def update_outcome():
     if request.method == "POST":
-        print("data =.",{
-        "session" : session["user_id"],
-        "nominal" : request.form["nominal"],
-        "used_type" : request.form["used_type"],
-        "note" : request.form["note"],
-        "date" : request.form["date"],
-        "id" : request.form["id"]
-    })
-        user_id = session["user_id"]
-        nominal = request.form["nominal"]
-        used_type = request.form["used_type"]
-        note = request.form["note"]
-        date = request.form["date"]
-        id = request.form["id"]
+        try:
+            user_id = session["user_id"]
+            nominal = request.form["nominal"]
+            used_type = request.form["used_type"]
+            note = request.form["note"]
+            date = request.form["date"]
+            id = request.form["id"]
 
-        outcome = Outcome(user_id=user_id, nominal=nominal, used_type=used_type, note=note, date=date, id=id)
-        outcome.update()
+            outcome = Outcome(user_id=user_id, nominal=nominal, used_type=used_type, note=note, date=date, id=id)
+            outcome.update()
 
-        flash('Data update successfully', 'success')
-        return redirect(url_for('outcome'))
+            flash('Data update successfully', 'success')
+            return redirect(url_for('outcome'))
+        except Exception as e:
+            print(f"Error update data: {e}")
+            flash('Error update data', 'error')
     return render_template('outcome.html')
 
 @app.route("/outcome/<outcome_id>/delete", methods=['GET', 'POST'])
 def delete_outcome(outcome_id):
     print(outcome_id)   
     if request.method == 'GET':
-        outcome = Outcome(None, None, None, None, None, None)
-        outcome.delete(outcome_id)
-        
-        flash('Data deleted successfully', 'success')
-        return redirect(url_for('outcome'))
+        try :
+            outcome = Outcome(None, None, None, None, None, None)
+            outcome.delete(outcome_id)
+            
+            flash('Data deleted successfully', 'success')
+            return redirect(url_for('outcome'))
+        except Exception as e:
+            print(f"Error deleted data: {e}")
+            flash('Error deleted data', 'error')
     return render_template('outcome.html')
 
 
@@ -236,9 +361,51 @@ def setting():
     cursor.execute("SELECT * FROM users WHERE id=%s", (user_id,))
     user = cursor.fetchone()
     dict_data = dict(zip([column[0] for column in cursor.description], user))
-    # return render_template('dashboard.html', data=dict_data)
     return render_template('setting.html', data=dict_data)
+@app.route("/setting/update", methods=['GET', 'POST'])
+def update_user():
+    if request.method == "POST":
+        try:
+            user_id = session["user_id"]
+            name = request.form["name"]
+            username = request.form["username"]
 
+            user = User(id=user_id, name=name,username=username)
+            user.update()
+
+            flash('Data update successfully', 'success')
+            return redirect(url_for('setting'))
+        except Exception as e:
+            print(f"Error update data: {e}")
+            flash('Error update data', 'error')
+    return render_template('setting.html')
+
+@app.route("/setting/update_password", methods=['GET', 'POST'])
+def update_password():
+    if request.method == "POST":
+        try:
+            user_id = session["user_id"]
+            current_password = request.form["current_password"]
+            new_password = request.form["new_password"]
+            cursor = conn.cursor()
+            cursor.execute("SELECT * FROM users WHERE id = %s", (user_id,))
+            user = cursor.fetchone()
+
+            if user and check_password_hash(user[3], current_password):
+                hashed_password = generate_password_hash(new_password, method='pbkdf2:sha256')
+                sql_update_income = """UPDATE users SET password=%s WHERE id=%s"""
+                cursor.execute(sql_update_income, (hashed_password, user_id))
+                conn.commit()
+                flash('Data update successfully', 'success')
+                # flash('Data update successfully', 'success')
+                return redirect(url_for('setting'))
+            else:      
+                flash('The current password is incorrect', 'error')
+                return redirect(url_for('setting'))
+        except Exception as e:
+            print(f"Error update data: {e}")
+            flash('Error update data', 'error')
+    return render_template('setting.html')
 @app.route('/register', methods=['GET', 'POST'])
 def register():
     if request.method == 'POST':
@@ -278,17 +445,17 @@ def login():
 
         if user and check_password_hash(user[3], password):
             session['user_id'] = user[0]
-            # flash('Login berhasil!', 'success')
+            session['name'] = user[1]
             return redirect(url_for('index'))
         else:
-            flash('Username atau password salah', 'danger')
+            flash('Incorrect username or password', 'error')
 
     return render_template('login.html')
 
 @app.route("/logout")
 def logout():
     session.pop("user_id", None)
-    flash('Logout berhasil!', 'success')
+    flash('Logout successful!', 'success')
     return redirect(url_for("login"))
